vscode config rework, and added config for payload debugging
Combined misc. configs and added a param for payload name. Now when your run a debug config, it simply asks you which payload (from tests) you want to use
Also, added payload debug (with gdb remote) - it works fine but sometimes
does not autostop on entry, so if that happens a simple breakpoint helps.

Note: there seems to be a bug in VSCode so if the payload starts infinite printing loop and you stop it via vscode debug 'pause' button, the debugged program stops but VScode goes into long and hard thinking. I usually end up just killing vscode. I'll leave it as is for now.