Refactor KM guest filesystem calls (version 2) (#149)
(There was a problem with the previous verion. It was easier to re-trace
my step to find the source of the problem.)

Anything using file descrptors or path names will need more processing
going forward. To support this I've refactored the system calls that
involve paths and fd into a separate file. Right now, the functionality
has moved into new functions (adding a new level of call). Next steps
are:
 - keep track of the file descriptors that the guest has open and
   ensure the guest doesn't use file descriptos it doesn't have open.
 - Support file prefixes for a poor man's chroot.