Change km_vcpu_init() to always create the gdb eventfd assigned to km_vcpu_t::gdb_efd so that it will be vaiid even when km is not run with the -g flag to use gdb. Also change km_vcpu_fini() to only close gdb_efd if it was opened. Changed km_vcpu_run() to log when it pauses on the eventfd gdb_efd.  This was used to discover when we are asked to pause when gdb was not be used with km.
This was tested by adding a new test to stray_test.c that would would create a new thread and then
the existing thread would execute an illegal instruction to cause the test to abort.
This sequence of events has the potential to cause km_vcpu_run() to be paused on the
gdb eventfd.  You just need to keep trying until it happens.  In my testing I found that
it almost always happened.  12 out of 13 tries would result in the new thread starting after the
old thread had executed the illegal instruction.
This test code is not being checked in since it does not seem to be a generally useful test.

Bug 209