The final set of changes to fix the gdb server entry race between threads that are breakpointing or signalling concurrently.
Add basic (but not complete) support for gdb remote protocol 'C' and 'S commands.
The gdb client needs to send signals back to the target using these commands.

Added a bats test named gdb_server_race that has two threads trying to wakeup the gdb server
thread concurrently.  One thread hits breakpoints.  The other thread has an illegal instruction
signal that fires continuously.  If the breakpoint and signal happen at or near the same time
the signal should always win so that we report that back to the gdd client.