fix mmap/mprotect + test
This fixes the crash in hello_2_loops_test with detached threads,
however it exposes some problem in C++ tls hadling. It seems
the tls destructor is called after the thread stack is unmapped.
Because of the bug in mmap/mprotect the stack was *not* really
unmapped so it masked the problem.