Add support for a gdb stop reply event queue to gdbstub.
This change lets gdbstub look at a queue of input events from payload threads that have tripped
breakpoints or received signals that cause the gdb client to become involved.
When a breakpoint fires or a thread receives a signal a gdb event is placed on a queue of pending
gdb events.  Then gdbstub is woken.  Then gdbstub looks at the queue and decides which event
to deliver to the gdb client in a stop reply.  Once the client has decided to let the km payload
run again gdbstub then needs to look at the event queue again to see if there are more events
that can be delivered before starting the payload.  If there is an event for a thread that will
be running, gdbstub delivers that event in a stop reply and goes back to waiting for input from
the gdb client.
The main intent of this fix is to ensure we detect when the gdb client is trying to step over
a planted breakpoint that has just been hit.  The gdb client will single step only that thread
past the breakpoint and then allow all the threads to run again.  If the gdb client receives
a stop reply for anything other than completion of that single step operation it will assert.
The gdb_server_race test has one thread hitting breakpoints continuously and another thread
executing an invalid instruction constantly.  So we did see cases where the SIGILL was delivered
while gdb client was trying to single step over a tripped breakpoint.  This change delays
delivery of inappropriate gdb events until an in progress breakpoint stepover operation completes.

While this change helps the gdb_server_race test pass more often on azure, I did find another
problem where the breakpointing thread managed to hit another breakpoint before a preceding
breakpoint event had been delivered to the gdb client.  This causes an assert in the gdbstub
to fire.