Add support for a gdb stop reply event queue to gdbstub.
This change lets gdbstub look at a queue of input events from payload threads that have tripped
breakpoints or received signals that cause the gdb client to become involved.
When a breakpoint fires or a thread receives a signal a gdb event is placed on a queue of pending
gdb events.  Then gdbstub is woken.  Then gdbstub looks at the queue and decides which event
to deliver to the gdb client in a stop reply.  Once the client has decided to let the km payload
run again gdbstub then needs to look at the event queue again to see if there are more events
that can be delivered before starting the payload.  If there is an event for a thread that will
be running, gdbstub delivers that event in a stop reply and goes back to waiting for input from
the gdb client.  In this case the target is never put into the running state.

This change ensures gdbstub doesn't deliver unexpected events to the gdb client.

Testing:
Ran the bats tests on a local workstation.
Run the bats tests on azure.  In addition run the gdb_server_race test many times on azure since
azure seems to preempt thread execution in ways that cause gdb stop replies to be deliver in
the wrong order.