Brought back running all tests as single BATS invocation (#390)
When refactoring tests, they were converted in up to 3 separate 'bats' runs
one for each type (e.g. so or dynamic). This created up to 3 different summaries and also
messed up ^C behavior to stop only a part of the test

This fix recovers the original behavior, while keeping single .bats source for so/dynamic/shared
Tested: Local runs with ^c , and CI