Address review comments and add support to gdbstub for libraries being loaded at runtime.
To get gdb to notice when a library is dynamically loaded it wants to place a breakpoint on dl_open_state().
This symbol is in the dynamic loader which gdb finds via the .interp section of the executable.
In .kmd executables .interp contains __km_dynlinker__ which can't be opened.
Changed the build to insert the proper path to the dynamic linker in the .interp section.
And changed some code in km to handle the .interp section containing the path to the musl
dynamic linker.

Fixup the damage done to tools/kontain-gcc by githooks precommit

And fix pre-commit to operate on files that end in .c, not just c.