Make /opt/kontain/runtime/libc.so the "production" home for musl libc.so
Now that payloads have the path of libc.so in their .interp section, we need
to have place for libc.so to live that is not in our development workspaces.
Change makefiles, and dockerfiles so that libc.so is placed in /opt/kontain/runtime
for both our dev machines and when running in containers.

Tested by running make and "make test" in my workspace.
Then by running "make withdocker" followed by "make test" to ensure that libc.so
from a docker build is available.
Then by building a CI (I think) test image and running the bats tests in kubernetes pod.

Fix dependency ordering in runtime/Makefile