Add basic gdb support for reading symbols from shared libraries. (#409)
* WIP Add basic gdb support for reading symbols from shared libraries.

Added the following gdb remote protocol commands:
 - qXfer:read:auxv
 - qXfer:libraries-svr4:read
 - qXfer:exec-file:read
 - vFile:

Added support for following the chain of link_map structures maintained by the dynamic linker
and dlopen().  Each link_map structure represents a loaded shared library.

Testing has been manual so far.  gdb does not automatically read symbols from the shared libraries.
The gdb "sharedlibrary" command must be run after the loader has run to coax gdb into reading the
shared library symbols.  Place a breakpoint at main(), then run to that breakpoint, then run the
sharedlibrary command.  Shared library symbols should be available after that.

* Add -G option to allow dynlinker to run and then stop at _start

Also added string_appen() which is a string append function that will grow the destination string if needed.
Address review comments.
Add code to remove all breakpoints when gdb detaches from the target.

* Move the code that waits for the dynamic linker to finish to gdb main loop.

This fixed the CI tests failures.

* Change verify_vcont() to allow a single thread running too.

So, now the rule is: allow 1 thread stepping, or allow 1 thread running, or allow all threads running.

* Fix the check for valid combinations of running and stepping threads in verify_vcont().

* bats test for the shared library related work.

* Address review comments and add support to gdbstub for libraries being loaded at runtime.

To get gdb to notice when a library is dynamically loaded it wants to place a breakpoint on dl_open_state().
This symbol is in the dynamic loader which gdb finds via the .interp section of the executable.
In .kmd executables .interp contains __km_dynlinker__ which can't be opened.
Changed the build to insert the proper path to the dynamic linker in the .interp section.
And changed some code in km to handle the .interp section containing the path to the musl
dynamic linker.

Fixup the damage done to tools/kontain-gcc by githooks precommit

And fix pre-commit to operate on files that end in .c, not just c.

* Add description to dlopen_exp.c and fixup pre-commit FILES list

* Make /opt/kontain/runtime/libc.so the "production" home for musl libc.so

Now that payloads have the path of libc.so in their .interp section, we need
to have place for libc.so to live that is not in our development workspaces.
Change makefiles, and dockerfiles so that libc.so is placed in /opt/kontain/runtime
for both our dev machines and when running in containers.

Tested by running make and "make test" in my workspace.
Then by running "make withdocker" followed by "make test" to ensure that libc.so
from a docker build is available.
Then by building a CI (I think) test image and running the bats tests in kubernetes pod.

Fix dependency ordering in runtime/Makefile