Change the gdb server in km to allow asynchronous gdb client attach at any time.
Change the km gdb server to always listen for connections from a gdb client and then
accept the connection.  The gdb client can then detach from km's gdb server and km
is still listening for connections from the gdb client so the client can reattach
later.

From the user perspective, the -g flag still causes km to pause the payload before it
starts up to allow a gdb client to attach.
If the -g flag is not supplied to km, the payload starts running and the km gdb
server is listening for connections from a gdb client.  A gdb client can attach
to km's gdb server at any time to debug the payload.

To test this a new bats test named gdb_attach was added.  This test has a payload
that does very little until it is instructed to stop.  While the km payload is wasting
time the test will attach the gdb client to km and print payload thread info and a stack
trace and then detach.  It does the same thing again to verify that the km gdb server
is still listening for gdb client attaches.  Then finally the test reattaches the gdb
client a final time to gracefully terminate the payload.