New 'flatten' target in demos/nokia, and small cleanup of demo image builds
Using redhat 'buildah' tool to flatten the image.

Ideally we should switch to buildah/skopeo/podman for docker work, but for now they can't work with Azure private Docker registry (some bug in auth), so we keep all as is and only use buildah to squash layers

It is not clear if it's good or bad for production, since flattening images makes lower layers content non-shareable, so this is more along the lines of preparations. We can use buildah to flattn original "FROM" images and still make them shareable; and generally we have lots of flexibility on how to construct layers now.