allow payload tests to interrogate KM internal vars, and use it in mmap test
Fixes #212

dropped UNITTEST hypercall - it was a hack to communicate
mmaps info back to payload for tests.
instead, introduced gdb python code that interrogates KM internal vars,
and is controlled by a 'write' content from the payload to a special fd
(-2020 - illegal when there is no related gdb code)
gdb breakpoints on reads, and for the special fd uses the buffer content
as a protocol to validate KM vars. That allows to write payload tests which
are aware of expected KM variables

gdb python code then makes decisions if the expected conditions are met,
and force payload write to return information back to payload .

Tests are included. More info in gdb/README.md