Change km command line options to allow disabling gdb server.
This changes the gdb related km command line options to the following:

-g[port]
-W[no|start|dynlink]

The presence of either -g or -W enables the gdb server inside of km.
If neither option is present the gdb server is not running and does not
listen for connections from a gdb client.

The -g option enables the km gdb server and optionally supplies the network
port the gdb server is to listen for gdb client connections on.
If no port is specified, the default port is 2159.  If -g is supplied but
-W is not, the the default value for -W is "start".

The -W option tells the gdb server what it should wait for before accepting
a gdb client connection.  The valid options are: no, start, dynlink.
If no option is supplied to -W, the default is "start".

no - the gdb server starts the payload immediately and waits for connections
   while the payload is running.

start - the gdb server runs the payload up to the _start entry point in
   the payload and then stops the payload and waits for a connection from
   the gdb client.  This is the default point to wait for gdb client attach.

dynlink - the gdb server does not run the payload and waits for gdb
   client connections before it runs the dynamic linker.

Once gdb server is enabled in km and successfuly listens for connections,
the gdb client may detach and reattach later.