Fix the demo with runenv images (#491)
* Fix demo under docker and k8s using the new runenv images.
* Fix missing libs for runenv-image python
* Created a runenv-demo-image target.
    * We want to seperate the demo with scripts and runenv-image that's bare minimum. It also serves as an example of how to consume runenv-image.
* Created NO_RUNENV variable. For components that doesn't have a concept of runenv, such as km and tests, this variable will remove the runenv related make targets.

Tested with demo-script.md. All steps works. Fix #480

With this change, km payload can be run on k8s the way we designed to be. runenv image will only carry the payload. KM will be installed on the host and be mounted in during runtime. custom yaml is still required, but the mechanism are all here.