Make the address field extraction code more consistent.
The shift then mask method makes clear to me how many of the least
significant bits are being discarded and then which bits of those remaining
are desired.
If we mask and then shift I need to mentally discard some of the mask too.
That is not too bad as long as the shift amount is a multiple of 4.