Change km command line options to allow disabling gdb server. (#490)
* Change km command line options to allow disabling gdb server.

This changes the gdb related km command line options to the following:

-g[port]
--gdb-listen
--gdb-dynlink

The presence of either -g, --gdb-listen, or --gdb-dynlink enables the gdb server inside of km.
If none of these options is present the gdb server is not running and does not
listen for connections from a gdb client.

The -g option enables the km gdb server and optionally supplies the network
port the gdb server is to listen for gdb client connections on.
If no port is specified, the default port is 2159.  If -g is supplied and neither
--gdb-listen or --gdb-dynlink is supplied then the gdb server waits at _start for the client to connect.

If --gdb-listen alone is supplied then the payload starts running and the gdb server listens for gdb client connections on port 2159.
If --gdb-dynlink alone is supplied then the gdb server listens for gdb client connections on port 2159 before the dynamic linker is run.

Once gdb server is enabled in km and successfully listens for connections,
the gdb client may detach and reattach later.