Fix id id: cannot find name for user ID 1001 (#561)
Inside the container, only uid is mapped, not the username. This is a known
issue with docker. For us, we invoke `make` inside the container, and some of
our make file includes calls `id -un` which prints the error. For now, removing
the offending variable will resolve the issue. However, this is only because we
can get away with not using the value.