Add the sigsuspend() hypercall to km.
nginx uses the sigsuspend() system call so we need to add it to km.
nginx also depends on reception of signals from outside of the process to
control nginx. So we need to be able to send signals to km and have those
signals forwarded in to the payload.
One of the signals nginx uses is SIGUSR1 which is the signal km uses to
interrupt ioctl( KVM_RUN ) calls used to execute payload instructions.
So, KM_SIGVCPUSTOP was redefined to SIGRTMAX-1 (gdb calls this SIG63).
This required some changes in km_core_tests.bats where SIGUSR1 is
coded in the gdb tests.
To test this the sigsuspend_test was added to the tests directory and
the sigsuspend bats test was added.

******Note that once this change is merged and pulled back to your work
******area, and if you have a .gdbinit file with the statement
******   "handle SIGUSR1 nostop",
******you will need to change SIGUSR1 to SIG63.

This was tested with the bats test and all tests passed.