Stop __km_syscall_handler from destroying rdx (#605)
Looks like the callers of syscalls expect at least rdx to be preserved.
__km_syscall_handler wasn't doing that which caused pthread_sigmask() to
seg fault which caused the sigsuspend test to wait forever for the begin
test flag file to exist.  This was fixed by restoring rdx from where it
was saved on the stack.  Also add code to restore the flags register.
Also enabled the sigsuspend test in native static tests.

Changed the sigsuspend_test to wait for 10 seconds for tje test to be ready to go.