Add fork() and clone() (for processes) support to km.
The fork() and clone() system calls are now handled by km.
In addition the wait4() and waitid() system calls have been added to allow
a parent process to wait for its children to terminate.

How it works:  When a thread forks (or clones) the parent process will write the the eventfd shutdown_fd
to transfer control to km's main() function where the fork or clone system call is performed.
All parent payload threads are paused for the fork/clone to complete.
After fork/clone the parent process resumes running as before the fork or clone.
The child process has only the km main thread running, all other threads don't exist.
So the child's km main rebuilds the vm and a single vcpu for the payload thread that should
be running following fork/clone and then sets that thread running.

km now keeps a table that maps between kontain pids and linux pids so that the payload only
sees kontain pids but km can supply the linux equivalent pid in calls to the linux kernel.
This pid mapping functionality is a stopgap at best.  There needs to be something better
to handle the conditions we will be seeing.

I've payed little attention to how gdb will be usable on the child km processes and their
payloads.  Additional followon worked is needed in this area too.

The kontain pid is now displayed in km_info() log messages since the child processes inherit
stdin, stdout, and stderr from the parent and we want to be able to tell which process is
generating which log messages.  We may need to revisit km's logging destination choices.

A new bats test verifies that fork(), clone(), wait4(), and waitid() are working.

All bats tests pass on my workstation.