Properly handle symlinks to km in payload name, as weel as in-shebang args
Usually shebang files (e.g. gunicorn) are parsed by the kernel,
but when shebang is passed to KM as a payload (e.g. on execve("gunincorn")), KM parses the shebang and extracts the payload name from there.
However, the payload name may point to something executable by KM (e.g. GO static binary) , or to a symlink to KM (e.g. "node").
We do not want KM to load elf of KM itself, so KM now does the same operation as it does with payload name - for payloads which are a symlink to KM,
the original payload name + .km is used

this commit also:

*  adds accurate handling of extra arg passed in shebang file on the first line
* removes a noisy trace
* fixes for fexecve and execveat