Add debug to the sigsuspend test.
Change the test to wait up to 5 seconds for the flagfile to come into existence,
in place of the current 1000 .01 second sleeps.
If the flagfile doesn't appear in 5 seconds, the test will dump the km trace for the test into
the TAP stream which should show up in the CI log.