Multiple fixes and clenup in python and symlinks
they were all done when chasing proper python invocation via symlinks and images cleanup, so they are bundles in the same PR

* Added handling of nested symlinks - we look for payload.km only at the last step, where symlinks points to KM
* Added python virtualenv cfg file so python selects correct libs without PYTHONPATH env
* added missing lib-ffl in Fedora buildenv - otherwise python configues to NOT use it and that excludes widely used _ctypes module
* corrected python images to use the above
* added sanity-check to python - validates we run in kontain and can load __ctypes which checks multiple steps in build
* made python container and on-the-box test use the same command

virtualenv cfg file info:

If python sees it' argv[0] as a symlink, it traverses to the target and tries to find
virtual env config file there. We use that by giving it virtenv file pointing back to the original cpython libs
It makes python->km approach work when KM and python are in different locations.

Tests included