Allow memory to be shared between parent and child payloads.
Changed the mmap_hcall() et al to mark an allocated piece of memory as shared if requested.
And when the memory is freed, change the memory back to private so it can be consolidated
into any neighboring pieces of free memory.

Wrote a test that creates a piece of shared memory, puts a semaphore in the shared memory,
forks a child process, and then have the parent and child operate on the shared memory using
the semaphore to synchronize access to the shared memory.

Change some bats tests to use mmap( MAP_PRIVATE ).
Now that MAP_SHARED is also supported some tests needed to either switch to using
MAP_PRIVATE or needed to specify what kind of memory is needed.

Tested using the bats tests.