Add handling for SIGCANCEL signal used by musl for pthread_cancel.
Also added symbols for all of the GDB_SIGNAL_* signal numbers in case we need
them in the future.

We ran into a problem debugging the pthread_cancel_test where the SIGCANCEL signal was
scrambled by gdb's signal number translation applied when the signal goes to the gdb client
and then again when the signal comes back to gdb stub.
SIGCANCEL was not in the table of signals that gdb stub knew how to translate so it when
through unchanged on the way to the client but did get changed when the signal came back.

All of the bats tests on a local workstation passed.

Tidy up the change for merge.