Change hypercalls to expect rax to point to km_hc_args_t
This replaces the former method of rax containing the lower 32 bits of the address of km_hc_args_t
And then resurrecting the upper bits inside of the hypercall dispatcher code.

To get this to work we have added this statement before making the ioctl( KVM_RUN ) call:
vcpu->cpu_run->kvm_valid_regs = KVM_SYNC_X86_REGS;
This tells kvm to return the vcpu's registers in the kvm_run structure when the ioctl returns.
Now that hypercall dispatch has the registers it can pickup rax directly and use that to access
the hypercall arguments.
All of the places in the guest where a hypercall (out %eax,(%rx)) is made were changed to set rax
with the address of the km_hc_args_t.  These changes were mostly in the c library.  And several others
were in the km code that is mapped into the guest (km/km_guest_asmcode.s).

Two hypercall test programs that make hyper calls with arguments but don't have much
real processing were added.  We are trying to test hypercall overhead with these.  One program is a single
thread, the other program has 10 threads doing the same as the single threaded program.  These
are not run by km_core_test.bats but we would like to keep them around.

This was tested by running the bats tests on a workstation.