Change Hypercall args to be stored in a per vcpu area addressed by the gs segment.
This change tries another way of storing hypercall arguments in a way that does not require rebuilding
the km_hc_args_t structure address from knowledge of the address of the base of the stack.  This
allows km to operate with the go langauge's co-routine method of multi-threading.
The basic change is an array of km_hc_args_t is stored in the part of km that is mapped into the guest's
address space and each vcpu's gs segment register is set to point to that vcpu's km_hc_args_t.
When a hypercall is made, the arguments are copied into the vcpu's private km_hc_args_t before making the
hypercall request with the x86 out instruction.
Since each vcpu has its own km_hc_args_t there is a case where a hypercall is interrupted by signal arrival,
then hc_ret is set for the returning hypercall.  Before returning to the payload, the signal handler is
entered and makes a hypercall which overwrites hc_ret from the interrupted hypercall.
We handle this by saving hc_ret in the signal handler's stack frame and restoring hc_ret when the signal
handler returns.

This was tested using the existing bats tests on a local workstation.