Faktory first implementation. (#711)
* faktory implementation.

split the image into application and base, then replace the base with our
implementation of base. only python is tried in this implementation.

* adds docker image definition

docker sdk has very difficult dependencies to work with,
so only copied nesserary APIs and struct definitions into
the repo.

* A helper tools to help understand the digest

* adds a python symlink in runenv-python

we need a python->python3 symblink, which is missing before.