Add a test to create 100 payload threads (#751)
* Add a test to create 100 payload threads

This test tries to expose problems that might appear when many payload threads are created.
This test was motivated by a problem where km_hcargs[] was created with a size that was too
small but large enough for most tests to pass.  When we finally went above 64 paylod threads
a problem happened.
The existing gdb_lost_of_threads_test was generalized to create a larger number of threads
up to the kvm maximum of KVM_MAX_VCPUS (288).

Tested with the new create-threads test in bats.

In addition all other bats tests are passing too.