Restore RSP in km_rt_sigreturn (#757)
* Restore RSP in km_rt_sigreturn

The Golang scheduler has a 'goroutine' preemption mechanism
that uses signals. Their signal handler will perform a context switch
by messing with the ucontext RIP and RSP.  Unfortunately we were
only restoring RIP (doh!).

* ucontext only on signal stack

We used to have 2 copies of the registers on the guest signal
stack, onle for ucontext and one for us. This was dumb.
Consolodate to ucontext. The side benefit is guest sigaction
can override all registers.

* save/restore rflags