Made payloads (node and python) images versioned  [trivial]
A part of prep for releasing versioned runenv images.
Added Major.minor to image names; the rest stays untouched.
This also makes it a little more reliable that we test the correct version.

Also added 'make release' placeholders, in case we want to release payloads as tarballs, not docker images

Tested by CI