Don't direct km tracing to stderr when it is a pipe or socket
Some applications need to know when a child process closes its stderr pipe or socket
that the parent is reading from.
When km dups stderr to KM_LOGGING this causes the file that fd 2 originally referred to be
be held open even if the payload closes fd 2.
This was causing the crun exec tests to fail when km was used to run the entrypoint of the
container.
We also found that the filesystem used by a running container could be read only so we
would not event be able to open a file to direct km tracing to.
km_redirect_msgs() was changed to set fd KM_LOGGING to an open file if the current fd 2
refers to a socket or a pipe.  If the file could not be opened, then km tracing output is
not saved.
To test this the km program --km-log-to flag accepts two keywords (stderr and none) that
forces logging to stderr even if it is a pipe or socket and none causes tracing to be lost.
The bats tests depend on km tracing being sent through a pipe so we needed to continue to
support this.

A new bats test named km_logging_test was written to exercise these options and verify
that they are doing what we want.