Use sigaction restorer instead of __km_sigreturn
It turns out that the libraies (libc and musl) pass the address of a
signal restorer function in 'struct sigaction'.

This change replaces the KM specific guest signal restorer
with the one that's already in the guest.

Oh yeah, and stack traces from inside signal handlers in GDB now works.