Add support for the getdents hypercall
Basically copied what we currently do for getdents64() and changed it to support getdents.
Added a getdents test to the filesys test.  Note that the test for getdents64 was
formerly called getdents.  Renamed that to getdents64 and the getdents test now
tests the older getdents hypercall.

Ran the bats tests to be sure nothing was broken by this change.