Eliminate the junk /tmp/km_$$.log files created by running km_core_tests.bats
Added code to open the km log file the first time km writes to it.

In km_core_tests.bats ensure that the .so related km command arguments appear before the payload name and the payload args

In km_core_tests.bats ensure that the snapshot file argument to a "km --resume" is placed at the end of the command since it is the payload name

Fixed various places in km_core_tests.bats where we run km without the use of km_with_timeout so that the --km-log-to=stderr arg is used.

When km resumes a snapshot that was derived from a .km.so payload, the resulting executable
is static and does not need the ldso related km command line options to resume.  So,
km_with_timeout() was modified to recognise the km --resume option and omit ldso km args
in this case.
In addition the bats snapshot tests were modified to place " -- " between the --resume argument
and the snapshot file name.  Since snapshot names are not usually recognised by km_with_timeout()
because they may not have a standard payload file extension we need to help km_with_timeout() by
using " -- ".

Tested by running the bats tests and then verifying that there are no file named /tmp/km_*.log