Fix the bats itimer test.
Our good buddy "azure the thread stall generator" has exposed a problem with the
bats itimer test.
The test blocks in sigsuspend() to await the signal that the interval timer has fired.
The timer is run in one shot mode so if the timer fires before the code gets to the call
to sigsuspend(), it waits until the test times out.  And then the test fails.
To fix this, the test now runs the timer in interval mode, so the timer keeps firing until
it is canceled at the end of the test.  That way if we suspend late, there will always be
another signal at the next interval expiration.

Tested by running the complete bats test suite.