Add support for gdb to follow fork() and execve() hypercalls. (#1022)
* Add support for gdb to follow fork() and execve() hypercalls.

How the gdb attach after fork and follow exec work...
Assuming gdb is enabled in km (using the -g flag)...
to have km wait for gdb attach in the child after a fork() call returns, set the KM_GDB_CHILD_FORK_WAIT env var in km's environment (not the payloads).
This env var will contain a regular expression.  If the name of the km payload matches the regular expression, km will wait for
the gdb client to attach to the child km.
To have the gdb client follow an execve() system call to the exec'ed program, run the gdb command "catch exec" before the execve()
call is made.  The gdb client will return to a prompt when notified that an exec has happened.
The payload will be stopped at _start.  You can start the payload using the gdb continue command.

Other changes made as part of this work:
-remove breakpoints from child process after fork
-update copyright date in touched source files
-add a global variable that contains the payload name
-break the vcpu start code out of km_start_vcpus() for fork
-add km gdb state to the env vars that km exec passes to the exec'ed program
-support KM_GDB_CHILD_FORK_WAIT km env var that controls whether km will wait for gdb client attach after returning from fork in the child process
-setup km_log_file after exec
-add gdb support for sending a gdb exec event to the gdb client after an exec succeeds
-km_gdb_setup_listen() now keeps lookinig for a free network if the one it tried is busy.  It only fails if no free port is found.
-km child processes inherit the parent's km command line gdb settings
-when unexpected interrupt is received km aborts now instead of exiting with non-zero return code.  This gives us a core file.
-removed the km_many bats test since km will keep looking for an available network port instead of disabling gdb when the requested port is unavailable
-add the gdb_forkexec bats test to test waiting for gdb client attach after a fork and to test following an exec to a new km payload
* Change km_redirect_msgs_after_exec() to handle no logging.
* Move the need pause all code into an inline function, reduce clutter