New HCALL for snapshot input and output data (#1082)
* New HCALL for snapshot input and output data

This change introduces support for snapshot Faas input and output data.
This allows data to be passed to and from a restarted snapshot. The
data is passed in files. New command line options (--input-data=
and --output-data=) allow data to be passed. A guest gets the data
with the HC_snapshot_getdata hypercall. Data is returned from the
guest with the HC_snapshot_putdata.

* fix file name error in test script