Change km pidmap to fail attempts to map invalid pids (#1106)
* Change km pidmap to fail attempts to map invalid pids

Issue #1100 reports a getsid(9999) system call succeeding when it should fail.
This is related to the pidmapping code in km returning the same pid it was passed
when it is unable to xlate from the passed process id to a kontain process id.
Since the translation failure if not detected the pid (9999) is passed in to the
linux kernel and occasionally 9999 would be a valid pid and getsid() would succeed.

This fix changes the km pidmap code to return a failure indication so km doesn't involve the
linux kernel when km knows that the request will fail.
And, km was silently allowed failed pid translations in both directions:
kontain -> linux
linux -> kontain
so both were changed to fail when the translation couldn't be done and the
calling code was changed to detect these failures and handle them.

In addition I found that km was not handling reception of a SIGCHLD signal after the
child process had already been reaped with wait4() or one of the other 16,000 variants of
waiting for a process.  So the pidmap was changed to retain entries for reaped processes
for the benefit of late arriving SIGCHLD signals.  The pidmap entries for reaped processes are now
used if no other free pidmap entries are found during allocation.

Tested with the bats tests.