Change km_post_signal() to interrupt a thread if possible (#1115)
* Change km_post_signal() to interrupt a thread if possible

There is a problem with some node tests not seeing the arrival of a SIGCHLD signal
when a child process terminates so the test just waits for the children.
The problem is in km_post_signal().  When a signal is sent to the process (as opposed to
a specific thread in the process) km_post_signal() does not interrupt any thread to have the
signal handler invoked.  If all threads are blocked waiting for something to happen and the
only thing that is going to happen is arrival of SIGCHLD, then the parent process is stuck.

The fix is to have km_post_signal() look for a thread to interrupt when a signal targeted to
the process arrives.  If a thread is found the system call the thread is blocking in is interrupted
so the signal handler can be called.

The fix currently only interrupts threads blocked in epoll_pwait() system calls.  This is the
system call the node tests were blocked in.
We could probably make this more general by targeting any thread blocking in a system call
that can return EINTR.

This was tested with "make -C payloads/node test-all" with a copy of payloads/node/skip_fork
that had fewer skipped tests in it and that passed.  Not including that version of skip_fork
hear since I don't understand how its contents where chosen.
In addition the bats tests are still passing.