Unblock SIGCHLD before starting the first vcpu in the child process after fork() (#1117)
* Unblock SIGCHLD before starting the first vcpu in the child.

We have a problem where a child of a fork() (or clone()) can't receive SIGCHLD when one of its
child processes terminate.  The cause of the problem is SIGCHLD (and SIGUSR1) are blocked when
the thread for the first vcpu is started.  The new thread inherits the creatin thread's signal mask.
The symptom is the child's child would terminate but wouldn't be wait()'ed on to reap the zombie child
process.

While trying to understand what was going on, I rearranged some code to make sure we get km tracing
going as soon as possible after exec happens.  In addition I found that the km exec state env vars
were still in the child's environment and were passed on to any children the child may fork and exec to.
So, then we had 2 copies of the exec recovery variables in the environment.  And, off course the first
set from the initial exec is what we found which caused at the least the child's child to have the wrong
pid.  To fix this we unset the exec recovery environment variables after we have completed exec recovery.

This was tested with the bats tests and some of the node tests.