Change exec hypercall to save and recover open file state needed by snapshots
The node payload tests surfaced some problems during execve() where fields in the km
open guest file table were not present because it was not transferred to the
exec target.

This change adds the needed information to contents of the KM_EXEC_GUESTFDS environment
variable used to pass this information from the execing program to the execed to program.

Some new traces and a new bats test program were added to verify that the open file state is
making it across the execve() hypercall.

The bats tests are passing on my workstation.
I still need to run the node tests to see if this change makes things better for it.